Refactor the ApplicationSettings to an interface
