+++ core/ext/src/commonTest/kotlin/com/chesire/nekomp/core/ext/TitlesExtTest.kt
@@ -1,7 +1,7 @@
 package com.chesire.nekomp.core.ext
 
 import com.chesire.nekomp.core.model.Titles
-import com.chesire.nekomp.core.preferences.TitleLanguage
+import com.chesire.nekomp.core.preferences.models.TitleLanguage
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
