+++ feature/library/src/commonMain/kotlin/com/chesire/nekomp/feature/library/ui/LibraryViewModel.kt
@@ -7,8 +7,8 @@ import com.chesire.nekomp.core.ext.toChosenLanguage
 import com.chesire.nekomp.core.model.EntryStatus
 import com.chesire.nekomp.core.model.Type
 import com.chesire.nekomp.core.preferences.ApplicationSettings
-import com.chesire.nekomp.core.preferences.ImageQuality
-import com.chesire.nekomp.core.preferences.TitleLanguage
+import com.chesire.nekomp.core.preferences.models.ImageQuality
+import com.chesire.nekomp.core.preferences.models.TitleLanguage
 import com.chesire.nekomp.feature.library.core.ObserveLibraryEntriesUseCase
 import com.chesire.nekomp.feature.library.core.RefreshLibraryEntriesUseCase
 import com.chesire.nekomp.feature.library.data.LibrarySettings
